[DT-4] CR fixes
