--- Spryker/Sniffs/DependencyProvider/AbstractDependencyProviderMethodAnnotationSniff.php
@@ -1,43 +1,0 @@
-<?php
-
-/**
- * MIT License
- * For full license information, please view the LICENSE file that was distributed with this source code.
- */
-
-namespace Spryker\Sniffs\DependencyProvider;
-
-use PHP_CodeSniffer\Files\File;
-use Spryker\Sniffs\AbstractSniffs\AbstractMethodAnnotationSniff;
-
-abstract class AbstractDependencyProviderMethodAnnotationSniff extends AbstractMethodAnnotationSniff
-{
-    /**
-     * @param \PHP_CodeSniffer\Files\File $phpCsFile
-     *
-     * @return bool
-     */
-    protected function isProvider(File $phpCsFile): bool
-    {
-        $className = $this->getClassName($phpCsFile);
-        $moduleName = $this->getModule($phpCsFile);
-
-        $providerName = $moduleName . 'DependencyProvider';
-        $stringLength = strlen($providerName);
-        $relevantClassNamePart = substr($className, -$stringLength);
-
-        return ($relevantClassNamePart === $providerName);
-    }
-
-    /**
-     * @param \PHP_CodeSniffer\Files\File $phpCsFile
-     *
-     * @return bool
-     */
-    protected function isCoreProvider(File $phpCsFile): bool
-    {
-        $namespace = $this->getNamespace($phpCsFile);
-
-        return ($namespace === static::NAMESPACE_SPRYKER);
-    }
-}
