--- Spryker/Sniffs/Facade/AbstractFacadeMethodAnnotationSniff.php
@@ -1,31 +1,0 @@
-<?php
-
-/**
- * MIT License
- * For full license information, please view the LICENSE file that was distributed with this source code.
- */
-
-namespace Spryker\Sniffs\Facade;
-
-use PHP_CodeSniffer\Files\File;
-use Spryker\Sniffs\AbstractSniffs\AbstractMethodAnnotationSniff;
-
-abstract class AbstractFacadeMethodAnnotationSniff extends AbstractMethodAnnotationSniff
-{
-    /**
-     * @param \PHP_CodeSniffer\Files\File $phpCsFile
-     *
-     * @return bool
-     */
-    protected function isFacade(File $phpCsFile): bool
-    {
-        $className = $this->getClassName($phpCsFile);
-        $moduleName = $this->getModule($phpCsFile);
-
-        $facadeName = $moduleName . 'Facade';
-        $stringLength = strlen($facadeName);
-        $relevantClassNamePart = substr($className, -$stringLength);
-
-        return ($relevantClassNamePart === $facadeName);
-    }
-}
